Small changes
i have modified the Start function, in order to fix the an error. When the $NEW_DOMAINS was null the command pihole -b was showing out an error message. It has been fixed changing the place where command was executed.

Now you can check all the new subdomains added when the script is installed.